+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 3.2.0
 
 -   Added `requestVerifiablePresentationV1` entrypoint to prove statements about identities and verifiable credentials with auditable anchor
+-   Added `ContractSchema` for `sendSponsoredTransaction` entrypoint for sponsored contract updates
 
 ### 3.1.0